Split `gdb-hs` into `Internal` and public modules
More involved wrappers can now more easily go into their own modules
